+++ packages/app/client/src/ui/editor/emulator/chatPanel/chatPanelContainer.ts
@@ -30,63 +30,21 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
-import * as React from 'react';
-import { shallow } from 'enzyme';
 
-import { Main } from './main';
+import { connect } from 'react-redux';
 
-function mockProxy() {
-  return new Proxy(
-    {},
-    {
-      get() {
-        return () => ({});
-      },
-    }
-  );
+import { RootState } from '../../../../state';
+
+import { ChatPanel, ChatPanelProps } from './chatPanel';
+
+function mapStateToProps(state: RootState, { documentId }: { documentId: string }): ChatPanelProps {
+  const endpointUrl = state.chat.chats[documentId] ? state.chat.chats[documentId].endpointUrl : '';
+  return {
+    endpointUrl,
+  };
 }
 
-jest.mock('electron', () => ({
-  remote: {
-    app: {
-      isPackaged: false,
-    },
-  },
-  ipcMain: new Proxy(
-    {},
-    {
-      get(): any {
-        return () => ({});
-      },
-      has() {
-        return true;
-      },
-    }
-  ),
-  ipcRenderer: new Proxy(
-    {},
-    {
-      get(): any {
-        return () => ({});
-      },
-      has() {
-        return true;
-      },
-    }
-  ),
-}));
-jest.mock('./explorer', () => mockProxy());
-jest.mock('./mdi', () => mockProxy());
-jest.mock('./navBar', () => mockProxy());
-jest.mock('./statusBar/statusBar.scss', () => ({}));
-jest.mock('../debug/storeVisualizer.scss', () => ({}));
-jest.mock('../../ui/dialogs', () => ({
-  DialogService: { showDialog: () => Promise.resolve(true) },
-}));
-
-describe('The Main component', () => {
-  it('should pass an empty test', () => {
-    const parent = shallow(<Main applicationMountComplete={() => void 0} />);
-    expect(parent.find(Main)).not.toBe(null);
-  });
-});
+export const ChatPanelContainer = connect(
+  mapStateToProps,
+  undefined
+)(ChatPanel);